[FSET-1119] GeneralApplicationReportingRepo reporting method moved to ReportingRepository. ReportingControllerSpec splitter by in several files, one per report.
