+++ app/repositories/application/GeneralApplicationRepository.scala
@@ -21,12 +21,10 @@ import java.util.regex.Pattern
 
 import common.Constants.{ No, Yes }
 import model.Adjustments._
-import model.ApplicationStatusOrder._
 import model.AssessmentScheduleCommands.{ ApplicationForAssessmentAllocation, ApplicationForAssessmentAllocationResult }
 import model.Commands._
 import model.EvaluationResults._
 import model.Exceptions.{ ApplicationNotFound, CannotUpdateReview, LocationPreferencesNotFound, SchemePreferencesNotFound }
-import common.StringUtils._
 import model.Exceptions._
 import model.PersistedObjects.ApplicationForNotification
 import model.Scheme.Scheme
@@ -36,7 +34,7 @@ import model.exchange.AssistanceDetails
 import org.joda.time.format.DateTimeFormat
 import org.joda.time.{ DateTime, LocalDate }
 import play.api.libs.json.{ Format, JsNumber, JsObject }
-import reactivemongo.api.{ DB, QueryOpts, ReadPreference }
+import reactivemongo.api.{ DB, QueryOpts }
 import reactivemongo.bson.{ BSONDocument, _ }
 import reactivemongo.json.collection.JSONBatchCommands.JSONCountCommand
 import repositories._
@@ -77,18 +75,6 @@ trait GeneralApplicationRepository {
 
   def updateQuestionnaireStatus(applicationId: String, sectionKey: String): Future[Unit]
 
-  def overallReport(frameworkId: String): Future[List[Report]]
-
-  def overallReportNotWithdrawn(frameworkId: String): Future[List[Report]]
-
-  def overallReportNotWithdrawnWithPersonalDetails(frameworkId: String): Future[List[ReportWithPersonalDetails]]
-
-  def adjustmentReport(frameworkId: String): Future[List[AdjustmentReport]]
-
-  def candidatesAwaitingAllocation(frameworkId: String): Future[List[CandidateAwaitingAllocation]]
-
-  def applicationsReport(frameworkId: String): Future[List[(String, IsNonSubmitted, PreferencesWithContactDetails)]]
-
   def confirmAdjustments(applicationId: String, data: Adjustments): Future[Unit]
 
   def findAdjustments(applicationId: String): Future[Option[Adjustments]]
@@ -102,12 +88,6 @@ trait GeneralApplicationRepository {
   def allocationExpireDateByApplicationId(applicationId: String): Future[Option[LocalDate]]
 
   def updateStatus(applicationId: String, status: ApplicationStatuses.EnumVal): Future[Unit]
-
-  def allApplicationAndUserIds(frameworkId: String): Future[List[PersonalDetailsAdded]]
-
-  def applicationsWithAssessmentScoresAccepted(frameworkId: String): Future[List[ApplicationPreferences]]
-
-  def applicationsPassedInAssessmentCentre(frameworkId: String): Future[List[ApplicationPreferencesWithTestResults]]
 
   def nextApplicationReadyForAssessmentScoreEvaluation(currentPassmarkVersion: String): Future[Option[String]]
 
@@ -419,513 +399,7 @@ class GeneralApplicationMongoRepository(timeZoneService: TimeZoneService)(implic
     collection.update(query, progressStatusBSON) map validator
   }
 
-  override def overallReportNotWithdrawn(frameworkId: String): Future[List[Report]] =
-    overallReport(BSONDocument("$and" -> BSONArray(
-      BSONDocument("frameworkId" -> frameworkId),
-      BSONDocument("applicationStatus" -> BSONDocument("$ne" -> ApplicationStatuses.Withdrawn))
-    )))
-
-  override def overallReportNotWithdrawnWithPersonalDetails(frameworkId: String): Future[List[ReportWithPersonalDetails]] =
-    overallReportWithPersonalDetails(BSONDocument("$and" -> BSONArray(
-      BSONDocument("frameworkId" -> frameworkId),
-      BSONDocument("applicationStatus" -> BSONDocument("$ne" -> ApplicationStatuses.Withdrawn))
-    )))
-
-  override def overallReport(frameworkId: String): Future[List[Report]] =
-    overallReport(BSONDocument("frameworkId" -> frameworkId))
-
-  private def overallReport(query: BSONDocument): Future[List[Report]] = {
-    val projection = BSONDocument(
-      "userId" -> "1",
-      "framework-preferences.alternatives.location" -> "1",
-      "framework-preferences.alternatives.framework" -> "1",
-      "framework-preferences.firstLocation.location" -> "1",
-      "framework-preferences.secondLocation.location" -> "1",
-      "framework-preferences.firstLocation.firstFramework" -> "1",
-      "framework-preferences.secondLocation.firstFramework" -> "1",
-      "framework-preferences.firstLocation.secondFramework" -> "1",
-      "framework-preferences.secondLocation.secondFramework" -> "1",
-      "personal-details.aLevel" -> "1",
-      "personal-details.stemLevel" -> "1",
-      "assistance-details.hasDisability" -> "1",
-      "assistance-details.needsSupportForOnlineAssessment" -> "1",
-      "assistance-details.needsSupportAtVenue" -> "1",
-      "assistance-details.guaranteedInterview" -> "1",
-      "issue" -> "1",
-      "applicationId" -> "1",
-      "progress-status" -> "2"
-    )
-
-    reportQueryWithProjections[BSONDocument](query, projection) map { lst =>
-      lst.map(docToReport)
-    }
-  }
-
-  override def applicationsWithAssessmentScoresAccepted(frameworkId: String): Future[List[ApplicationPreferences]] =
-    applicationPreferences(BSONDocument("$and" -> BSONArray(
-      BSONDocument("frameworkId" -> frameworkId),
-      BSONDocument(s"progress-status.${ProgressStatuses.AssessmentScoresAcceptedProgress}" -> true),
-      BSONDocument("applicationStatus" -> BSONDocument("$ne" -> ApplicationStatuses.Withdrawn))
-    )))
-
-  // scalastyle:off method.length
-  private def applicationPreferences(query: BSONDocument): Future[List[ApplicationPreferences]] = {
-    val projection = BSONDocument(
-      "userId" -> "1",
-      "framework-preferences.alternatives.location" -> "1",
-      "framework-preferences.alternatives.framework" -> "1",
-      "framework-preferences.firstLocation.location" -> "1",
-      "framework-preferences.secondLocation.location" -> "1",
-      "framework-preferences.firstLocation.firstFramework" -> "1",
-      "framework-preferences.secondLocation.firstFramework" -> "1",
-      "framework-preferences.firstLocation.secondFramework" -> "1",
-      "framework-preferences.secondLocation.secondFramework" -> "1",
-      "assistance-details.hasDisability" -> "1",
-      "assistance-details.needsSupportForOnlineAssessment" -> 1,
-      "assistance-details.needsSupportAtVenue" -> 1,
-      "assistance-details.guaranteedInterview" -> "1",
-      "personal-details.aLevel" -> "1",
-      "personal-details.stemLevel" -> "1",
-      "passmarkEvaluation" -> "2",
-      "applicationId" -> "1"
-    )
-
-    reportQueryWithProjections[BSONDocument](query, projection).map { list =>
-      list.map { document =>
-        val userId = document.getAs[String]("userId").getOrElse("")
-
-        val fr = document.getAs[BSONDocument]("framework-preferences")
-
-        val fr1 = fr.flatMap(_.getAs[BSONDocument]("firstLocation"))
-        val fr1FirstLocation = fr1.flatMap(_.getAs[String]("location"))
-        val fr1FirstFramework = fr1.flatMap(_.getAs[String]("firstFramework"))
-        val fr1SecondFramework = fr1.flatMap(_.getAs[String]("secondFramework"))
-
-        val fr2 = fr.flatMap(_.getAs[BSONDocument]("secondLocation"))
-        val fr2FirstLocation = fr2.flatMap(_.getAs[String]("location"))
-        val fr2FirstFramework = fr2.flatMap(_.getAs[String]("firstFramework"))
-        val fr2SecondFramework = fr2.flatMap(_.getAs[String]("secondFramework"))
-
-        val frAlternatives = fr.flatMap(_.getAs[BSONDocument]("alternatives"))
-        val location = frAlternatives.flatMap(_.getAs[Boolean]("location").map(booleanTranslator))
-        val framework = frAlternatives.flatMap(_.getAs[Boolean]("framework").map(booleanTranslator))
-
-        val ad = document.getAs[BSONDocument]("assistance-details")
-        val hasDisability = ad.flatMap(_.getAs[String]("hasDisability"))
-        val needsSupportForOnlineAssessment = ad.flatMap(_.getAs[Boolean]("needsSupportForOnlineAssessment"))
-        val needsSupportAtVenue = ad.flatMap(_.getAs[Boolean]("needsSupportAtVenue"))
-        // TODO: Revisit this and check if we want to return both fields. If we want to keep one decide what to do if both values are None
-        val needsAdjustment: Option[String] = (needsSupportForOnlineAssessment, needsSupportAtVenue) match {
-          case (None, None) => None
-          case (Some(true), _) => Some("Yes")
-          case (_, Some(true)) => Some("Yes")
-          case _ => Some("No")
-        }
-        val guaranteedInterview = ad.flatMap(_.getAs[Boolean]("guaranteedInterview")).map { gis => if (gis) "Yes" else "No"}
-
-        val pd = document.getAs[BSONDocument]("personal-details")
-        val aLevel = pd.flatMap(_.getAs[Boolean]("aLevel").map(booleanTranslator))
-        val stemLevel = pd.flatMap(_.getAs[Boolean]("stemLevel").map(booleanTranslator))
-
-        val applicationId = document.getAs[String]("applicationId").getOrElse("")
-
-        val pe = document.getAs[BSONDocument]("passmarkEvaluation")
-
-        val otLocation1Scheme1PassmarkEvaluation = pe.flatMap(_.getAs[String]("location1Scheme1").map(Result(_).toPassmark))
-        val otLocation1Scheme2PassmarkEvaluation = pe.flatMap(_.getAs[String]("location1Scheme2").map(Result(_).toPassmark))
-        val otLocation2Scheme1PassmarkEvaluation = pe.flatMap(_.getAs[String]("location2Scheme1").map(Result(_).toPassmark))
-        val otLocation2Scheme2PassmarkEvaluation = pe.flatMap(_.getAs[String]("location2Scheme2").map(Result(_).toPassmark))
-        val otAlternativeSchemePassmarkEvaluation = pe.flatMap(_.getAs[String]("alternativeScheme").map(Result(_).toPassmark))
-
-        ApplicationPreferences(userId, applicationId, fr1FirstLocation, fr1FirstFramework, fr1SecondFramework,
-          fr2FirstLocation, fr2FirstFramework, fr2SecondFramework, location, framework, hasDisability,
-          guaranteedInterview, needsAdjustment, aLevel, stemLevel,
-          OnlineTestPassmarkEvaluationSchemes(otLocation1Scheme1PassmarkEvaluation, otLocation1Scheme2PassmarkEvaluation,
-            otLocation2Scheme1PassmarkEvaluation, otLocation2Scheme2PassmarkEvaluation, otAlternativeSchemePassmarkEvaluation))
-      }
-    }
-  }
-
-  // scalstyle:on method.length
-
-  override def applicationsPassedInAssessmentCentre(frameworkId: String): Future[List[ApplicationPreferencesWithTestResults]] =
-    applicationPreferencesWithTestResults(BSONDocument("$and" -> BSONArray(
-      BSONDocument("frameworkId" -> frameworkId),
-      BSONDocument(s"progress-status.${ProgressStatuses.AssessmentCentrePassedProgress}" -> true),
-      BSONDocument("applicationStatus" -> BSONDocument("$ne" -> ApplicationStatuses.Withdrawn))
-    )))
-
-  // scalastyle:off method.length
-  private def applicationPreferencesWithTestResults(query: BSONDocument): Future[List[ApplicationPreferencesWithTestResults]] = {
-    val projection = BSONDocument(
-      "userId" -> "1",
-      "framework-preferences.alternatives.location" -> "1",
-      "framework-preferences.alternatives.framework" -> "1",
-      "framework-preferences.firstLocation.location" -> "1",
-      "framework-preferences.secondLocation.location" -> "1",
-      "framework-preferences.firstLocation.firstFramework" -> "1",
-      "framework-preferences.secondLocation.firstFramework" -> "1",
-      "framework-preferences.firstLocation.secondFramework" -> "1",
-      "framework-preferences.secondLocation.secondFramework" -> "1",
-      "assessment-centre-passmark-evaluation" -> "2",
-      "applicationId" -> "1",
-      "personal-details.firstName" -> "1",
-      "personal-details.lastName" -> "1",
-      "personal-details.preferredName" -> "1",
-      "personal-details.aLevel" -> "1",
-      "personal-details.stemLevel" -> "1",
-      "personal-details.dateOfBirth" -> "1"
-    )
-
-    reportQueryWithProjections[BSONDocument](query, projection).map { list =>
-      list.map { document =>
-        val userId = document.getAs[String]("userId").getOrElse("")
-
-        val fr = document.getAs[BSONDocument]("framework-preferences")
-
-        val fr1 = fr.flatMap(_.getAs[BSONDocument]("firstLocation"))
-        val fr1FirstLocation = fr1.flatMap(_.getAs[String]("location"))
-        val fr1FirstFramework = fr1.flatMap(_.getAs[String]("firstFramework"))
-        val fr1SecondFramework = fr1.flatMap(_.getAs[String]("secondFramework"))
-
-        val fr2 = fr.flatMap(_.getAs[BSONDocument]("secondLocation"))
-        val fr2FirstLocation = fr2.flatMap(_.getAs[String]("location"))
-        val fr2FirstFramework = fr2.flatMap(_.getAs[String]("firstFramework"))
-        val fr2SecondFramework = fr2.flatMap(_.getAs[String]("secondFramework"))
-
-        val frAlternatives = fr.flatMap(_.getAs[BSONDocument]("alternatives"))
-        val location = frAlternatives.flatMap(_.getAs[Boolean]("location").map(booleanTranslator))
-        val framework = frAlternatives.flatMap(_.getAs[Boolean]("framework").map(booleanTranslator))
-
-        val applicationId = document.getAs[String]("applicationId").getOrElse("")
-
-        val pe = document.getAs[BSONDocument]("assessment-centre-passmark-evaluation")
-
-        val ca = pe.flatMap(_.getAs[BSONDocument]("competency-average"))
-        val leadingAndCommunicatingAverage = ca.flatMap(_.getAs[Double]("leadingAndCommunicatingAverage"))
-        val collaboratingAndPartneringAverage = ca.flatMap(_.getAs[Double]("collaboratingAndPartneringAverage"))
-        val deliveringAtPaceAverage = ca.flatMap(_.getAs[Double]("deliveringAtPaceAverage"))
-        val makingEffectiveDecisionsAverage = ca.flatMap(_.getAs[Double]("makingEffectiveDecisionsAverage"))
-        val changingAndImprovingAverage = ca.flatMap(_.getAs[Double]("changingAndImprovingAverage"))
-        val buildingCapabilityForAllAverage = ca.flatMap(_.getAs[Double]("buildingCapabilityForAllAverage"))
-        val motivationFitAverage = ca.flatMap(_.getAs[Double]("motivationFitAverage"))
-        val overallScore = ca.flatMap(_.getAs[Double]("overallScore"))
-
-        val se = pe.flatMap(_.getAs[BSONDocument]("schemes-evaluation"))
-        val commercial = se.flatMap(_.getAs[String](Schemes.Commercial).map(Result(_).toPassmark))
-        val digitalAndTechnology = se.flatMap(_.getAs[String](Schemes.DigitalAndTechnology).map(Result(_).toPassmark))
-        val business = se.flatMap(_.getAs[String](Schemes.Business).map(Result(_).toPassmark))
-        val projectDelivery = se.flatMap(_.getAs[String](Schemes.ProjectDelivery).map(Result(_).toPassmark))
-        val finance = se.flatMap(_.getAs[String](Schemes.Finance).map(Result(_).toPassmark))
-
-        val pd = document.getAs[BSONDocument]("personal-details")
-        val firstName = pd.flatMap(_.getAs[String]("firstName"))
-        val lastName = pd.flatMap(_.getAs[String]("lastName"))
-        val preferredName = pd.flatMap(_.getAs[String]("preferredName"))
-        val aLevel = pd.flatMap(_.getAs[Boolean]("aLevel").map(booleanTranslator))
-        val stemLevel = pd.flatMap(_.getAs[Boolean]("stemLevel").map(booleanTranslator))
-        val dateOfBirth = pd.flatMap(_.getAs[LocalDate]("dateOfBirth"))
-
-        ApplicationPreferencesWithTestResults(userId, applicationId, fr1FirstLocation, fr1FirstFramework,
-          fr1SecondFramework, fr2FirstLocation, fr2FirstFramework, fr2SecondFramework, location, framework,
-          PersonalInfo(firstName, lastName, preferredName, aLevel, stemLevel, dateOfBirth),
-          CandidateScoresSummary(leadingAndCommunicatingAverage, collaboratingAndPartneringAverage,
-            deliveringAtPaceAverage, makingEffectiveDecisionsAverage, changingAndImprovingAverage,
-            buildingCapabilityForAllAverage, motivationFitAverage, overallScore),
-          SchemeEvaluation(commercial, digitalAndTechnology, business, projectDelivery, finance))
-      }
-    }
-  }
-
-  // scalstyle:on method.length
-
-  private def overallReportWithPersonalDetails(query: BSONDocument): Future[List[ReportWithPersonalDetails]] = {
-    val projection = BSONDocument(
-      "userId" -> "1",
-      "framework-preferences.alternatives.location" -> "1",
-      "framework-preferences.alternatives.framework" -> "1",
-      "framework-preferences.firstLocation.location" -> "1",
-      "framework-preferences.secondLocation.location" -> "1",
-      "framework-preferences.firstLocation.firstFramework" -> "1",
-      "framework-preferences.secondLocation.firstFramework" -> "1",
-      "framework-preferences.firstLocation.secondFramework" -> "1",
-      "framework-preferences.secondLocation.secondFramework" -> "1",
-      "personal-details.aLevel" -> "1",
-      "personal-details.dateOfBirth" -> "1",
-      "personal-details.firstName" -> "1",
-      "personal-details.lastName" -> "1",
-      "personal-details.preferredName" -> "1",
-      "personal-details.stemLevel" -> "1",
-      "online-tests.cubiksUserId" -> "1",
-      "assistance-details.hasDisability" -> "1",
-      "assistance-details.needsSupportForOnlineAssessment" -> "1",
-      "assistance-details.needsSupportAtVenue" -> "1",
-      "assistance-details.guaranteedInterview" -> "1",
-      "applicationId" -> "1",
-      "progress-status" -> "2"
-    )
-
-    reportQueryWithProjections[BSONDocument](query, projection) map { lst =>
-      lst.map(docToReportWithPersonalDetails)
-    }
-  }
-
-  private def docToReport(document: BSONDocument) = {
-    val fr = document.getAs[BSONDocument]("framework-preferences")
-    val fr1 = fr.flatMap(_.getAs[BSONDocument]("firstLocation"))
-    val fr2 = fr.flatMap(_.getAs[BSONDocument]("secondLocation"))
-
-    def frLocation(root: Option[BSONDocument]) = extract("location")(root)
-
-    def frScheme1(root: Option[BSONDocument]) = extract("firstFramework")(root)
-
-    def frScheme2(root: Option[BSONDocument]) = extract("secondFramework")(root)
-
-    val personalDetails = document.getAs[BSONDocument]("personal-details")
-    val aLevel = personalDetails.flatMap(_.getAs[Boolean]("aLevel").map(booleanTranslator))
-    val stemLevel = personalDetails.flatMap(_.getAs[Boolean]("stemLevel").map(booleanTranslator))
-
-    val fpAlternatives = fr.flatMap(_.getAs[BSONDocument]("alternatives"))
-    val location = fpAlternatives.flatMap(_.getAs[Boolean]("location").map(booleanTranslator))
-    val framework = fpAlternatives.flatMap(_.getAs[Boolean]("framework").map(booleanTranslator))
-
-    val applicationId = document.getAs[String]("applicationId").getOrElse("")
-    val progress: ProgressResponse = findProgress(document, applicationId)
-
-    val issue = document.getAs[String]("issue")
-
-    val ad = document.getAs[BSONDocument]("assistance-details")
-    val hasDisability = ad.flatMap(_.getAs[String]("hasDisability"))
-    val needsSupportForOnlineAssessment = ad.flatMap(_.getAs[Boolean]("needsSupportForOnlineAssessment"))
-    val needsSupportAtVenue = ad.flatMap(_.getAs[Boolean]("needsSupportAtVenue"))
-    // TODO: Revisit this and check if we want to return both fields. If we want to keep one decide what to do if both values are None
-    val needsAdjustments: Option[String] = (needsSupportForOnlineAssessment, needsSupportAtVenue) match {
-      case (None, None) => None
-      case (Some(true), _) => Some("Yes")
-      case (_, Some(true)) => Some("Yes")
-      case _ => Some("No")
-    }
-    val guaranteedInterview = ad.flatMap(_.getAs[Boolean]("guaranteedInterview")).map { gis => if (gis) Yes else No}
-
-    Report(
-      applicationId, Some(getStatus(progress)), frLocation(fr1), frScheme1(fr1), frScheme2(fr1),
-      frLocation(fr2), frScheme1(fr2), frScheme2(fr2), aLevel,
-      stemLevel, location, framework, hasDisability, needsAdjustments, guaranteedInterview, issue
-    )
-  }
-
-  private def docToReportWithPersonalDetails(document: BSONDocument) = {
-    val fr = document.getAs[BSONDocument]("framework-preferences")
-    val fr1 = fr.flatMap(_.getAs[BSONDocument]("firstLocation"))
-    val fr2 = fr.flatMap(_.getAs[BSONDocument]("secondLocation"))
-
-    def frLocation(root: Option[BSONDocument]) = extract("location")(root)
-
-    def frScheme1(root: Option[BSONDocument]) = extract("firstFramework")(root)
-
-    def frScheme2(root: Option[BSONDocument]) = extract("secondFramework")(root)
-
-    val personalDetails = document.getAs[BSONDocument]("personal-details")
-    val aLevel = personalDetails.flatMap(_.getAs[Boolean]("aLevel").map(booleanTranslator))
-    val stemLevel = personalDetails.flatMap(_.getAs[Boolean]("stemLevel").map(booleanTranslator))
-    val firstName = personalDetails.flatMap(_.getAs[String]("firstName"))
-    val lastName = personalDetails.flatMap(_.getAs[String]("lastName"))
-    val preferredName = personalDetails.flatMap(_.getAs[String]("preferredName"))
-    val dateOfBirth = personalDetails.flatMap(_.getAs[String]("dateOfBirth"))
-
-    val fpAlternatives = fr.flatMap(_.getAs[BSONDocument]("alternatives"))
-    val location = fpAlternatives.flatMap(_.getAs[Boolean]("location").map(booleanTranslator))
-    val framework = fpAlternatives.flatMap(_.getAs[Boolean]("framework").map(booleanTranslator))
-
-    val ad = document.getAs[BSONDocument]("assistance-details")
-    val hasDisability = ad.flatMap(_.getAs[String]("hasDisability"))
-    val needsSupportForOnlineAssessment = ad.flatMap(_.getAs[Boolean]("needsSupportForOnlineAssessment"))
-    val needsSupportAtVenue = ad.flatMap(_.getAs[Boolean]("needsSupportAtVenue"))
-    // TODO: Revisit this and check if we want to return both fields. If we want to keep one decide what to do if both values are None
-    val needsAdjustment: Option[String] = (needsSupportForOnlineAssessment, needsSupportAtVenue) match {
-      case (None, None) => None
-      case (Some(true), _) => Some(Yes)
-      case (_, Some(true)) => Some(Yes)
-      case _ => Some(No)
-    }
-    val guaranteedInterview = ad.flatMap(_.getAs[Boolean]("guaranteedInterview")).map { gis => if (gis) Yes else No }
-
-    val applicationId = document.getAs[String]("applicationId").getOrElse("")
-    val userId = document.getAs[String]("userId").getOrElse("")
-    val progress: ProgressResponse = findProgress(document, applicationId)
-
-    val onlineTests = document.getAs[BSONDocument]("online-tests")
-    val cubiksUserId = onlineTests.flatMap(_.getAs[Int]("cubiksUserId"))
-
-    ReportWithPersonalDetails(
-      applicationId, userId, Some(getStatus(progress)), frLocation(fr1), frScheme1(fr1), frScheme2(fr1),
-      frLocation(fr2), frScheme1(fr2), frScheme2(fr2), aLevel,
-      stemLevel, location, framework, hasDisability, needsAdjustment, guaranteedInterview, firstName, lastName,
-      preferredName, dateOfBirth, cubiksUserId
-    )
-  }
-
-  def adjustmentReport(frameworkId: String): Future[List[AdjustmentReport]] = {
-    val query = BSONDocument("$and" ->
-      BSONArray(
-        BSONDocument("frameworkId" -> frameworkId),
-        BSONDocument("applicationStatus" -> BSONDocument("$ne" -> ApplicationStatuses.InProgress)),
-        BSONDocument("applicationStatus" -> BSONDocument("$ne" -> ApplicationStatuses.Withdrawn)),
-        BSONDocument("$or" ->
-          BSONArray(
-            BSONDocument("$or" -> BSONArray(
-              BSONDocument("assistance-details.needsSupportForOnlineAssessment" -> true),
-              BSONDocument("assistance-details.needsSupportAtVenue" -> true))),
-            BSONDocument("assistance-details.guaranteedInterview" -> true)
-          ))
-      ))
-
-    val projection = BSONDocument(
-      "userId" -> "1",
-      "personal-details.firstName" -> "1",
-      "personal-details.lastName" -> "1",
-      "personal-details.preferredName" -> "1",
-      "assistance-details.guaranteedInterview" -> "1",
-      "assistance-details.typeOfAdjustments" -> "1",
-      "assistance-details.needsSupportForOnlineAssessment" -> "1",
-      "assistance-details.needsSupportAtVenue" -> "1",
-      "assistance-details.adjustmentsConfirmed" -> "1"
-    )
-
-    reportQueryWithProjections[BSONDocument](query, projection).map { list =>
-      list.map { document =>
-
-        val personalDetails = document.getAs[BSONDocument]("personal-details")
-        val userId = document.getAs[String]("userId").getOrElse("")
-        val firstName = extract("firstName")(personalDetails)
-        val lastName = extract("lastName")(personalDetails)
-        val preferredName = extract("preferredName")(personalDetails)
-
-        val ad = document.getAs[BSONDocument]("assistance-details")
-        val guaranteedInterview = ad.flatMap(_.getAs[Boolean]("guaranteedInterview")).map { gis => if (gis) Yes else No }
-        val adjustmentsConfirmed = getAdjustmentsConfirmed(ad)
-        val typesOfAdjustments = ad.flatMap(_.getAs[List[String]]("typeOfAdjustments"))
-        val adjustments = typesOfAdjustments.getOrElse(Nil)
-        val finalTOA = if (adjustments.isEmpty) None else Some(adjustments.map(splitCamelCase).mkString("|"))
-
-        AdjustmentReport(userId, firstName, lastName, preferredName, None, None, finalTOA, guaranteedInterview, adjustmentsConfirmed)
-      }
-    }
-  }
-
-  def candidatesAwaitingAllocation(frameworkId: String): Future[List[CandidateAwaitingAllocation]] = {
-    val query = BSONDocument("$and" ->
-      BSONArray(
-        BSONDocument("frameworkId" -> frameworkId),
-        BSONDocument("applicationStatus" -> ApplicationStatuses.AwaitingAllocation)
-      ))
-
-    val projection = BSONDocument(
-      "userId" -> "1",
-      "personal-details.firstName" -> "1",
-      "personal-details.lastName" -> "1",
-      "personal-details.preferredName" -> "1",
-      "personal-details.dateOfBirth" -> "1",
-      "framework-preferences.firstLocation.location" -> "1",
-      "assistance-details.typeOfAdjustments" -> "1",
-      "assistance-details.needsSupportForOnlineAssessment" -> "1",
-      "assistance-details.needsSupportAtVenue" -> "1"
-    )
-
-    reportQueryWithProjections[BSONDocument](query, projection).map { list =>
-      list.map { document =>
-
-        val userId = document.getAs[String]("userId").get
-        val personalDetails = document.getAs[BSONDocument]("personal-details").get
-        val firstName = personalDetails.getAs[String]("firstName").get
-        val lastName = personalDetails.getAs[String]("lastName").get
-        val preferredName = personalDetails.getAs[String]("preferredName").get
-        val dateOfBirth = personalDetails.getAs[LocalDate]("dateOfBirth").get
-        val frameworkPreferences = document.getAs[BSONDocument]("framework-preferences").get
-        val firstLocationDoc = frameworkPreferences.getAs[BSONDocument]("firstLocation").get
-        val firstLocation = firstLocationDoc.getAs[String]("location").get
-
-        val assistance = document.getAs[BSONDocument]("assistance-details")
-        val typesOfAdjustments = assistance.flatMap(_.getAs[List[String]]("typeOfAdjustments"))
-
-        val adjustments = typesOfAdjustments.getOrElse(Nil)
-        val finalTOA = if (adjustments.isEmpty) None else Some(adjustments.map(splitCamelCase).mkString("|"))
-
-        CandidateAwaitingAllocation(userId, firstName, lastName, preferredName, firstLocation, finalTOA, dateOfBirth)
-      }
-    }
-  }
-
-  def applicationsReport(frameworkId: String): Future[List[(String, IsNonSubmitted, PreferencesWithContactDetails)]] = {
-    val query = BSONDocument("frameworkId" -> frameworkId)
-
-    val projection = BSONDocument(
-      "applicationId" -> "1",
-      "personal-details.preferredName" -> "1",
-      "userId" -> "1",
-      "framework-preferences" -> "1",
-      "progress-status" -> "2"
-    )
-
-    val seed = Future.successful(List.empty[(String, Boolean, PreferencesWithContactDetails)])
-    reportQueryWithProjections[BSONDocument](query, projection).flatMap { lst =>
-      lst.foldLeft(seed) { (applicationsFuture, document) =>
-        applicationsFuture.map { applications =>
-          val timeCreated = isoTimeToPrettyDateTime(getDocumentId(document).time)
-          val applicationId = document.getAs[String]("applicationId").get
-          val personalDetails = document.getAs[BSONDocument]("personal-details")
-          val preferredName = extract("preferredName")(personalDetails)
-          val userId = document.getAs[String]("userId").get
-          val frameworkPreferences = document.getAs[Preferences]("framework-preferences")
-
-          val location1 = frameworkPreferences.map(_.firstLocation.location)
-          val location1Scheme1 = frameworkPreferences.map(_.firstLocation.firstFramework)
-          val location1Scheme2 = frameworkPreferences.flatMap(_.firstLocation.secondFramework)
-
-          val location2 = frameworkPreferences.flatMap(_.secondLocation.map(_.location))
-          val location2Scheme1 = frameworkPreferences.flatMap(_.secondLocation.map(_.firstFramework))
-          val location2Scheme2 = frameworkPreferences.flatMap(_.secondLocation.flatMap(_.secondFramework))
-
-          val p = findProgress(document, applicationId)
-
-          val preferences = PreferencesWithContactDetails(None, None, preferredName, None, None,
-            location1, location1Scheme1, location1Scheme2,
-            location2, location2Scheme1, location2Scheme2,
-            Some(ApplicationStatusOrder.getStatus(p)), Some(timeCreated))
-
-          (userId, isNonSubmittedStatus(p), preferences) +: applications
-        }
-      }
-    }
-  }
-
-  private def getDocumentId(document: BSONDocument): BSONObjectID =
-    document.get("_id").get match {
-      case id: BSONObjectID => id
-      case id: BSONString => BSONObjectID(id.value)
-    }
-
-  private def isoTimeToPrettyDateTime(utcMillis: Long): String =
-    timeZoneService.localize(utcMillis).toString("yyyy-MM-dd HH:mm:ss")
-
-  private def booleanTranslator(bool: Boolean) = if (bool) Yes else No
-
-  private def reportQueryWithProjections[A](
-                                             query: BSONDocument,
-                                             prj: BSONDocument,
-                                             upTo: Int = Int.MaxValue,
-                                             stopOnError: Boolean = true
-                                           )(implicit reader: Format[A]): Future[List[A]] =
-    collection.find(query).projection(prj).cursor[A](ReadPreference.nearest).collect[List](upTo, stopOnError)
-
   def extract(key: String)(root: Option[BSONDocument]) = root.flatMap(_.getAs[String](key))
-
-  private def getAdjustmentsConfirmed(assistance: Option[BSONDocument]): Option[String] = {
-    assistance.flatMap(_.getAs[Boolean]("adjustmentsConfirmed")).getOrElse(false) match {
-      case false => Some("Unconfirmed")
-      case true => Some("Confirmed")
-    }
-  }
 
   def confirmAdjustments(applicationId: String, data: Adjustments): Future[Unit] = {
 
@@ -1041,22 +515,6 @@ class GeneralApplicationMongoRepository(timeZoneService: TimeZoneService)(implic
     ))
 
     collection.update(query, applicationStatusBSON, upsert = false) map { _ => () }
-  }
-
-  def allApplicationAndUserIds(frameworkId: String): Future[List[PersonalDetailsAdded]] = {
-    val query = BSONDocument("frameworkId" -> frameworkId)
-    val projection = BSONDocument(
-      "applicationId" -> "1",
-      "userId" -> "1"
-    )
-
-    collection.find(query, projection).cursor[BSONDocument]().collect[List]().map {
-      _.map { doc =>
-        val userId = doc.getAs[String]("userId").getOrElse("")
-        val applicationId = doc.getAs[String]("applicationId").getOrElse("")
-        PersonalDetailsAdded(applicationId, userId)
-      }
-    }
   }
 
   def nextApplicationReadyForAssessmentScoreEvaluation(currentPassmarkVersion: String): Future[Option[String]] = {
